i don't really want to make a separate commit for all this

+++ src/main/java/org/codeorange/frc2024/subsystem/wrist/Wrist.java
@@ -37,7 +37,7 @@ public class Wrist extends AbstractSubsystem {
         wristIO.zeroWristEncoder();
     }
 
-    public void runVoltage(int volts) {
+    public void runVoltage(double volts) {
         wristIO.setVoltage(volts);
     }
 }
